am facut: - criptare peste tot - rezolvare problema aliniere design - schimbare dimensiune mesaje - compatibilitate path cu / sau \ (imi pun baza in cum interpreteaza java)

+++ client/src/main/java/andrei/dynamic/client/MessageFromServer.java
@@ -16,7 +16,7 @@ public class MessageFromServer {
 
     }
 
-    public MessageFromServer(byte type, byte[] content) {
+    public MessageFromServer(int type, byte[] content) {
 
 	this.type = MessageType.parseCode(type);
 	this.content = content;
